attention-experiment/requirements.txt
backcall
chardet
decorator
einops
fire
greenlet
idna
inflection
ipdb
ipython
ipython-genutils
jedi
msgpack
munch
neovim
nltk
numpy
parso
pexpect
pickleshare
Pillow
prompt-toolkit
ptyprocess
Pygments
pynvim
pytorch-ignite
six
tensorboard
tensorboardX
#tensorflow-gpu
torch
torchtext
torchvision
tqdm
traitlets
urllib3
wcwidth
rouge
unidecode
absl-py
#tensorflow
#tensorflow-datasets
#tensorflow-hub
pytorch-lightning
test_tube
trains
matplotlib
wandb
scipy
scikit-learn
opencv-python
coverage
codecov
pytest
pytest-cov
pytest-flake8
flake8
flake8-black
check-manifest
twine
isort
pre-commit
mypy
yapf
matplotlib
scipy
h5py
pandas
wfdb
pytorch-lightning-bolts
ConSSL
keras
efficientnet_pytorch
